remove react-router-dom from components directory
