+++ components/src/index.ts
@@ -18,7 +18,6 @@ export * from './lists'
 export * from './modals'
 export * from './nav'
 export * from './primitives'
-export * from './tabbedNav'
 export * from './slotmap'
 export * from './structure'
 export * from './tooltips'
