+++ components/src/lists/index.ts
@@ -1,4 +1,3 @@
 // list and list item components
 export * from './SidePanelGroup'
 export * from './TitledList'
-export * from './ListItem'
