--- components/src/tabbedNav/index.ts
@@ -1,7 +1,0 @@
-// navigational components
-
-// TODO(bc, 2021-03-29): these components are only used in one place
-// reconsider whether they belong in components library
-export * from './TabbedNavBar'
-export * from './NavTab'
-export * from './OutsideLinkTab'
