+++ protocol-designer/src/containers/NavTab.tsx
@@ -1,21 +1,23 @@
 import * as React from 'react'
-import cx from 'classnames'
+import { NavLink } from 'react-router-dom'
+import classnames from 'classnames'
+
+import { Button, NotificationIcon } from '@opentrons/components'
+import type { IconName, ButtonProps } from '@opentrons/components'
 
 import styles from './navbar.module.css'
-import { Button } from '../buttons'
-import { NotificationIcon } from '../icons'
 
-import type { IconName } from '../icons'
-
-export interface OutsideLinkTabProps {
+export interface NavTabProps {
   /** optional click event for nav button */
   onClick?: React.MouseEventHandler
-  /** link to outside URL */
-  to: string
+  /** optional url for nav button route */
+  url?: string
   /** position a single button on the bottom of the page */
   isBottom?: boolean
   /** classes to apply */
   className?: string
+  /** id */
+  id?: string
   /** disabled attribute (setting disabled removes onClick) */
   disabled?: boolean
   /** optional title to display below the icon */
@@ -28,23 +30,32 @@ export interface OutsideLinkTabProps {
   selected?: boolean
 }
 
-/** Very much like NavTab, but used for opening external links in a new tab/window */
-export function OutsideLinkTab(props: OutsideLinkTabProps): JSX.Element {
-  const className = cx(props.className, styles.tab, styles.no_link, {
+export function NavTab(props: NavTabProps): JSX.Element {
+  const { url } = props
+  const className = classnames(props.className, styles.tab, {
     [styles.disabled]: props.disabled,
     [styles.bottom]: props.isBottom,
     [styles.selected]: props.selected,
   })
+
+  let buttonProps: ButtonProps = {
+    id: props.id,
+    className: className,
+    disabled: props.disabled,
+    onClick: props.onClick,
+  }
+
+  if (url) {
+    buttonProps = {
+      ...buttonProps,
+      Component: NavLink,
+      to: url,
+      activeClassName: styles.selected,
+    }
+  }
+
   return (
-    <Button
-      className={className}
-      disabled={props.disabled}
-      onClick={props.onClick}
-      Component="a"
-      href={props.disabled ? '' : props.to}
-      target="_blank"
-      rel="noopener noreferrer"
-    >
+    <Button {...buttonProps}>
       <NotificationIcon
         name={props.iconName}
         childName={props.notification ? 'circle' : null}
