+++ protocol-designer/src/containers/OutsideLinkTab.tsx
@@ -1,25 +1,19 @@
 import * as React from 'react'
-import { NavLink } from 'react-router-dom'
-import classnames from 'classnames'
+import cx from 'classnames'
+import { Button, NotificationIcon } from '@opentrons/components'
+import type { IconName } from '@opentrons/components'
 
 import styles from './navbar.module.css'
-import { Button } from '../buttons'
-import { NotificationIcon } from '../icons'
 
-import type { IconName } from '../icons'
-import type { ButtonProps } from '../buttons'
-
-export interface NavTabProps {
+export interface OutsideLinkTabProps {
   /** optional click event for nav button */
   onClick?: React.MouseEventHandler
-  /** optional url for nav button route */
-  url?: string
+  /** link to outside URL */
+  to: string
   /** position a single button on the bottom of the page */
   isBottom?: boolean
   /** classes to apply */
   className?: string
-  /** id */
-  id?: string
   /** disabled attribute (setting disabled removes onClick) */
   disabled?: boolean
   /** optional title to display below the icon */
@@ -32,32 +26,23 @@ export interface NavTabProps {
   selected?: boolean
 }
 
-export function NavTab(props: NavTabProps): JSX.Element {
-  const { url } = props
-  const className = classnames(props.className, styles.tab, {
+/** Very much like NavTab, but used for opening external links in a new tab/window */
+export function OutsideLinkTab(props: OutsideLinkTabProps): JSX.Element {
+  const className = cx(props.className, styles.tab, styles.no_link, {
     [styles.disabled]: props.disabled,
     [styles.bottom]: props.isBottom,
     [styles.selected]: props.selected,
   })
-
-  let buttonProps: ButtonProps = {
-    id: props.id,
-    className: className,
-    disabled: props.disabled,
-    onClick: props.onClick,
-  }
-
-  if (url) {
-    buttonProps = {
-      ...buttonProps,
-      Component: NavLink,
-      to: url,
-      activeClassName: styles.selected,
-    }
-  }
-
   return (
-    <Button {...buttonProps}>
+    <Button
+      className={className}
+      disabled={props.disabled}
+      onClick={props.onClick}
+      Component="a"
+      href={props.disabled ? '' : props.to}
+      target="_blank"
+      rel="noopener noreferrer"
+    >
       <NotificationIcon
         name={props.iconName}
         childName={props.notification ? 'circle' : null}
